backend/requirements.txt
amqp==2.6.1
asgiref==3.2.10
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.7
coverage==5.2.1
Django==3.0.8
django-filter==2.3.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
kombu==4.6.11
Markdown==3.2.2
phonenumbers==8.12.7
Pillow==7.1.1
pkg-resources==0.0.0
pytz==2020.1
PyYAML==5.3.1
pyyml==0.0.2
sqlparse==0.3.1
vine==1.3.0
